Fix ReadFileData() signed/unsigned issues with byte values >= 2 GiB.
The primary beneficiary of these changes is GetVAT().
But in the course of these modifications, it was discovered that that 
function was written against the UDF 1.50 specification; UDF 2.00
saw a significant rework of the VAT implementation.

+++ udf/tools/src/chkudf/protos.h
@@ -291,8 +291,9 @@ int ReadSectors(void *buffer, UINT32 address, UINT32 Count);
 
 int ReadLBlocks(void *buffer, UINT32 address, UINT16 partition, UINT32 Count);
 
-int ReadFileData(void *buffer, const struct FE_or_EFE *ICB, UINT16 part,
-                 int startOffset, int bytesRequested, UINT32 *data_start_loc);
+unsigned int ReadFileData(void *buffer, const struct FE_or_EFE *ICB, UINT16 part,
+                          unsigned long long startOffset, unsigned int bytesRequested,
+                          UINT32 *data_start_loc);
 
 /*****************************************************************************
  * verifyAVDP.c
